+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-30 10:34:07
-// Commit: feat(test): integrate custom telemetry logger in test
+// Updated at: 2026-07-01 15:31:17
+// Commit: feat(test): implement shimmer loading skeleton in test
 
 class testComponentState {
-  final String commitRef = '2421';
-  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
-  final DateTime timestamp = DateTime.parse('2026-06-30T10:34:07+05:00');
+  final String commitRef = '2450';
+  final String lastMessage = 'feat(test): implement shimmer loading skeleton in test';
+  final DateTime timestamp = DateTime.parse('2026-07-01T15:31:17+05:00');
 }